Modularization and fixes for legacy header files and std:: prefixes
