+++ .gitignore
@@ -16,3 +16,4 @@ Release
 Generated Files
 obj
 vsix/LICENSE
+_build
