+++ strings/base_abi.h
@@ -1,5 +1,5 @@
 
-namespace winrt::impl
+WINRT_EXPORT namespace winrt::impl
 {
     template <> struct abi<Windows::Foundation::IUnknown>
     {
