+++ strings/base_collections.h
@@ -1,5 +1,5 @@
 
-namespace winrt::impl
+WINRT_EXPORT namespace winrt::impl
 {
     namespace wfc = Windows::Foundation::Collections;
