+++ strings/base_collections_input_map_view.h
@@ -1,5 +1,5 @@
 
-namespace winrt::impl
+WINRT_EXPORT namespace winrt::impl
 {
     template <typename K, typename V, typename Container>
     struct input_map_view :
