+++ strings/base_collections_input_vector.h
@@ -1,5 +1,5 @@
 
-namespace winrt::impl
+WINRT_EXPORT namespace winrt::impl
 {
     template <typename T, typename Container, typename ThreadingBase>
     struct vector_impl :
