+++ strings/base_coroutine_threadpool.h
@@ -1,5 +1,5 @@
 
-namespace winrt::impl
+WINRT_EXPORT namespace winrt::impl
 {
     inline auto submit_threadpool_callback(void(__stdcall* callback)(void*, void* context), void* context)
     {
@@ -319,7 +319,7 @@ WINRT_EXPORT namespace winrt
     };
 }
 
-namespace winrt::impl
+WINRT_EXPORT namespace winrt::impl
 {
     struct apartment_awaiter
     {
