+++ strings/base_delegate.h
@@ -1,5 +1,5 @@
 
-namespace winrt::impl
+WINRT_EXPORT namespace winrt::impl
 {
 #if defined(_MSC_VER) && !defined(__clang__)
 #pragma warning(push)
