+++ strings/base_fast_forward.h
@@ -28,7 +28,7 @@ static_assert(WINRT_FAST_ABI_SIZE >= %);
 
 #pragma detect_mismatch("WINRT_FAST_ABI_SIZE", WINRT_IMPL_STRING(WINRT_FAST_ABI_SIZE))
 
-namespace winrt::impl
+WINRT_EXPORT namespace winrt::impl
 {
     // Thunk definitions are in arch-specific assembly sources
 %
