+++ strings/base_marshaler.h
@@ -1,5 +1,5 @@
 
-namespace winrt::impl
+WINRT_EXPORT namespace winrt::impl
 {
     inline std::int32_t make_marshaler(unknown_abi* outer, void** result) noexcept
     {
