+++ strings/base_meta.h
@@ -48,7 +48,7 @@ WINRT_EXPORT namespace winrt
     }
 }
 
-namespace winrt::impl
+WINRT_EXPORT namespace winrt::impl
 {
     using namespace std::literals;
