+++ strings/base_reference_produce.h
@@ -1,5 +1,5 @@
 
-namespace winrt::impl
+WINRT_EXPORT namespace winrt::impl
 {
     template <typename T>
     struct reference : implements<reference<T>, Windows::Foundation::IReference<T>, Windows::Foundation::IPropertyValue>
@@ -420,7 +420,7 @@ WINRT_EXPORT namespace winrt::Windows::Foundation
     }
 }
 
-namespace winrt::impl
+WINRT_EXPORT namespace winrt::impl
 {
     template <typename T, typename From>
     T unbox_value_type(From&& value)
