+++ strings/base_std_hash.h
@@ -1,5 +1,5 @@
 
-namespace winrt::impl
+WINRT_EXPORT namespace winrt::impl
 {
     inline std::size_t hash_data(void const* ptr, std::size_t const bytes) noexcept
     {
