+++ strings/base_string_input.h
@@ -30,7 +30,7 @@ WINRT_EXPORT namespace winrt::param
 
         hstring(wchar_t const* const value) noexcept
         {
-            create_string_reference(value, wcslen(value));
+            create_string_reference(value, std::wcslen(value));
         }
 
         operator winrt::hstring const&() const noexcept
@@ -41,7 +41,7 @@ WINRT_EXPORT namespace winrt::param
     private:
         void create_string_reference(wchar_t const* const data, std::size_t size) noexcept
         {
-            WINRT_ASSERT(size < UINT_MAX);
+            WINRT_ASSERT(size < (std::numeric_limits<std::uint32_t>::max)());
             auto size32 = static_cast<std::uint32_t>(size);
 
             if (size32 == 0)
@@ -65,7 +65,7 @@ WINRT_EXPORT namespace winrt::param
     }
 }
 
-namespace winrt::impl
+WINRT_EXPORT namespace winrt::impl
 {
     template <typename T>
     using param_type = std::conditional_t<std::is_same_v<T, hstring>, param::hstring, T>;
