+++ strings/base_windows.h
@@ -1,5 +1,5 @@
 
-namespace winrt::impl
+WINRT_EXPORT namespace winrt::impl
 {
 #ifdef WINRT_DIAGNOSTICS
